more debugging

+++ main.py
@@ -26,8 +26,8 @@ class Snake:
 class Food:
     def __init__(self):
         
-        x = random.randint(0, (GAME_WIDTH/SPACE_SIZE) - 1) * SPACE_SIZE
-        y = random.randint(0, (GAME_HEIGHT/SPACE_SIZE) - 1) * SPACE_SIZE
+        x = float(random.randint(0, (GAME_WIDTH/SPACE_SIZE) - 1) * SPACE_SIZE)
+        y = float(random.randint(0, (GAME_HEIGHT/SPACE_SIZE) - 1) * SPACE_SIZE)
         
         self.coordinates = [x, y]
